Moving Month extension methods to separate class, and simplifying some method names
